+++ bee-message/src/output/nft.rs
@@ -10,7 +10,7 @@ use crate::{
 use bee_common::packable::{Packable, Read, Write};
 
 ///
-const METADATA_LENGTH_MAX: u32 = 1024;
+const METADATA_LENGTH_MAX: usize = 1024;
 
 ///
 pub struct NftOutputBuilder {
@@ -61,8 +61,8 @@ impl NftOutputBuilder {
 
     ///
     pub fn finish(self) -> Result<NftOutput, Error> {
-        if self.immutable_metadata.len() > METADATA_LENGTH_MAX as usize {
-            return Err(Error::InvalidMetadataLength(self.immutable_metadata.len() as u32));
+        if self.immutable_metadata.len() > METADATA_LENGTH_MAX {
+            return Err(Error::InvalidMetadataLength(self.immutable_metadata.len()));
         }
 
         Ok(NftOutput {
@@ -163,11 +163,11 @@ impl Packable for NftOutput {
         let amount = u64::unpack_inner::<R, CHECK>(reader)?;
         let native_tokens = NativeTokens::unpack_inner::<R, CHECK>(reader)?;
         let nft_id = NftId::unpack_inner::<R, CHECK>(reader)?;
-        let immutable_metadata_len = u32::unpack_inner::<R, CHECK>(reader)?;
+        let immutable_metadata_len = u32::unpack_inner::<R, CHECK>(reader)? as usize;
         if CHECK && immutable_metadata_len > METADATA_LENGTH_MAX {
             return Err(Error::InvalidMetadataLength(immutable_metadata_len));
         }
-        let mut immutable_metadata = vec![0u8; immutable_metadata_len as usize];
+        let mut immutable_metadata = vec![0u8; immutable_metadata_len];
         reader.read_exact(&mut immutable_metadata)?;
         let feature_blocks = FeatureBlocks::unpack_inner::<R, CHECK>(reader)?;
